feat(tooling): corda AIO image + connector + tests
Adds the all in one Corda image for convenient tests
and examples that can be self contained without having
to pull up a separate ledger prior to launching cactus
tests or examples.

The AIO image uses the Braid RPC server to provide
communication between the Cactus Corda ledger connector
and the Corda ledger itself. This means that if you
want to deploy Cactus to interact with a Corda ledger
of yours you *will* need to have a Braid server running
there as well.
For further details on Braid RPC server see the Cordite
Gitlab repo:
https://gitlab.com/cordite/braid

Because #143 and #280 ended up being closely dependent
on each other it has been decided that this commit will
solve the both of them in a single swoop since otherwise
it would be just extra work to implement the non-Braid
solution first only to have it immediately superseded by
the Braid based solution provided as part of #280.

Fixes #143

Fixes #280

Partial credit here goes to @jweate who started the work
on this branch.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>